Move overkiz cover definitions
