+++ homeassistant/components/overkiz/cover/generic_cover.py
