--- homeassistant/components/overkiz/cover_entities/__init__.py
@@ -1,1 +1,0 @@
-"""Cover entities for the Overkiz (by Somfy) integration."""
